+++ compose/ui/ui/src/skikoMain/kotlin/androidx/compose/ui/scene/ComposeSceneInputHandler.skiko.kt
@@ -114,6 +114,7 @@ internal class ComposeSceneInputHandler(
             button,
         )
         prepareForPointerInputEvent()
+        updatePointerPosition()
         syntheticEventSender.send(event)
         updatePointerPositions(event)
     }
@@ -123,7 +124,7 @@ internal class ComposeSceneInputHandler(
         return processKeyEvent(keyEvent)
     }
 
-    fun sendSyntheticInputEvents() {
+    fun updatePointerPosition() {
         syntheticEventSender.updatePointerPosition()
     }
